+++ packages/components/components/elvis-chip/src/react/elvia-chip.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useState, useEffect } from 'react';
+import React, { FC, useState, useEffect, CSSProperties } from 'react';
 import { ChipCheckmarkIcon, ChipComponent, ChipCloseIcon, ChipTitle } from './styledComponents';
 import { ChipType, ColorType, onChangeValue } from './elvia-chip.types';
 
@@ -13,6 +13,8 @@ export interface BaseChipProps {
   value: string;
   onDelete?: (event: string) => void;
   valueOnChange?: (event: onChangeValue) => void;
+  className?: string;
+  inlineStyle?: { [style: string]: CSSProperties };
   webcomponent?: any;
 }
 
@@ -25,6 +27,8 @@ export const Chip: FC<BaseChipProps> = ({
   value,
   onDelete,
   valueOnChange,
+  className,
+  inlineStyle,
   webcomponent,
 }) => {
   const [isSelected, setIsSelected] = useState(selected);
@@ -69,6 +73,8 @@ export const Chip: FC<BaseChipProps> = ({
       onMouseLeave={setHover(false)}
       chipType={type}
       isSelected={isSelected}
+      className={`${className ? className : ''}`}
+      style={inlineStyle}
       data-testid="chip-button"
     >
       {type === 'choice' && (