+++ packages/components/components/elvis-progress-linear/src/react/elvia-progress-linear.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React, { CSSProperties, FC } from 'react';
 import './style.scss';
 import classnames from 'classnames';
 
@@ -6,9 +6,17 @@ export interface ProgressbarProps {
   value: number;
   isIndeterminate: boolean;
   isError: boolean;
+  className?: string;
+  inlineStyle?: { [style: string]: CSSProperties };
 }
 
-const ProgressLinear: React.FC<ProgressbarProps> = ({ value = 0, isIndeterminate, isError }) => {
+const ProgressLinear: FC<ProgressbarProps> = ({
+  value = 0,
+  isIndeterminate,
+  isError,
+  className,
+  inlineStyle,
+}) => {
   const classes = classnames({
     ['ewc-progress-linear--range']: !isIndeterminate && !isError,
     ['ewc-progress-linear--indeterminate']: isIndeterminate && !isError,
@@ -16,8 +24,10 @@ const ProgressLinear: React.FC<ProgressbarProps> = ({ value = 0, isIndeterminate
   });
 
   return (
-    <div className="ewc-progress-linear">
-      <div className={classes} style={{ width: `${value}%` }} data-testid="progress-linear"></div>
+    <div className={className ? className : ''} style={inlineStyle} data-testid="progress-wrapper">
+      <div className={'ewc-progress-linear'}>
+        <div className={classes} style={{ width: `${value}%` }} data-testid="progress-linear"></div>
+      </div>
     </div>
   );
 };