+++ packages/components/components/elvis-card/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Card Changelog
 
+## 1.1.0 (07.02.22)
+
+### New feature
+
+- Added inlineStyle and className props
+
 ## 1.0.1 (07.01.22)
 
 ### New feature