+++ packages/components/components/elvis-icon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Icon Changelog
 
+## 1.1.0 (07.02.22)
+
+### New feature
+
+- Added inlineStyle and className props
+
 ## 1.0.1 (04.01.22))
 
 Bump to make npm package public